refactor: support using the arguments in run_managed
This commit changes `AppCommand.run_managed` from a property to a method
that receives the cli's parsed arguments and should still return whether
the command should run in managed mode or not.

There is no change in functionality; this refactoring is meant as a
starting point for future commits. For instance, consider that the
"pack" command can run in managed more or not, depending on whether the
"--destructive-mode" flag is used (once that feature is supported).