feat: fetch logs from inside the build instance (#87)
* feat: fetch logs from inside the build instance

One could argue that this a bugfix since this was missing behavior from
the existing craft tools. Regardless, this commit updates the instance()
method in the default ProviderService to fetch the logfile from inside
the build instance once its done. The contents of this file are then
emitted so that they can end up in the logs of the "outer" instance.

Fixes #59