fix: merge command groups with the same name
The use-case here is where the application provides commands in a
logical "group" with the same name as of the default ones (a common
example is an "init" command in the "Other" group).

Instead of having multiple command groups with the same name, which
works fine but looks ugly when printing the help, merge command groups
with the same name into a single one.

Fixes #94