feat: do variable expansion when loading project (#101)
Perform environment expansion on the yaml dict that is used to create
the Project. This expands variables like CRAFT_PROJECT_VERSION (and the
many more provided by craft-parts).

Also add a "hook" for app-specific modifications to the yaml dict in the
form of the `_extra_yaml_transform()` method (a no-op in the base
Application).