chore: refactor "features"
Instead of having a "loose" boolean in Application, formalize the
concept a bit by adding an AppFeatures object to be kept in the
AppMetadata (as part of the app's "static" info).

This complicates testing a bit because the features have to be enabled
at AppMetadata-creation-time and can't be set "a posteriori" anymore,
but the end result isn't too onerous and better reflects what the
applications will do anyway.