feat: provide a hook to customize validation errors
This commit adds a new transform_pydantic_errors() classmethod on
CraftBaseModel to let subclasses customize/convert/translate
Pydantic-generated error messages for fields that are specific to those
subclasses.

This new "hook" is then used to update the error messages for the "name"
and "version" fields of the Project class: instead of showing the raw
"regex", we copy over Snapcraft's nice and user-friendly messages that
fully describe what is and isn't allowed.

One caveat is that Pydantic 1.x's errors don't include the original
(user-provided) value, so we can't use it on the message. We can fix
this once we move to Pydantic 2.