fix(remote-build): set a known path to cacerts
This is a (manual) cherry-pick of commit 375553d6fc15 from snapcraft,
with minor updates to make the comments more generic.

This gist of the fix is that pygit2 tries to initialize certificate locations
at import-time, and when the application is running as a snap the certs are
located in a non-standard (core-based) directory.