feat(lifecycle)!: don't build on incompatible bases
The Lifecycle service will now check whether the build's base (as stored
in the BuildInfo) matches the running environment. This doesn't affect
managed runs (barring bugs), but is a change in behavior as we will no
longer build, say, Ubuntu 22.04 artifacts on Fedora hosts in destructive
mode.

We define "matches" here as this:

- If the build base is "devel", then only the systems have to match. For
  instance, you can build a "devel" Ubuntu-based artefact on any Ubuntu
  system, but not on a Fedora system.
- Otherwise, both the base's system and version must match, such that it
  is no longer possible to build ubuntu@22.04 on (say) ubuntu@20.04 systems.

Fixes #154