feat: start the fetch-service (#8)
Add a set of utility functions so that FetchService.setup() will spawn a
fetch-service if necessary. These utility functions include a function to
get the service's (json) status, and a function to tell whether the service
is up.

Use the $SNAP_USER_COMMON location for the fetch-service snap as a base for
the 'config' and 'spool' subdirectories.