feat: configure build instances for fetch-service (#11)
This configuration happens when creating the session. Currently we:

- Install the local certificate (added to this repo)
- Configure pip, snapd and apt
- Refresh apt's package listings to ensure all installations are traceable
  by the fetch-service

This initial implementation *always* refreshes apt listings, even if the
project build won't use them (e.g. no stage- or build- packages). In the
future we'll be smarter about this. The implementation also only supports
LXD instances, just because we need to figure out the instance's gateway
address and only the LXD work for that has been done so far. In the future
we can support all instance types with a cleaner craft-providers API.