feat: generate the fetch-service certificate
Recent versions of the fetch-service not longer come with a fixed, "built-in"
certificate. Instead, clients are expected to generate their own. To do this,
we use a location that is shared among all Craft applications - the key and
certificate are generated if not found, and are otherwise shared by all
instances of all applications.

Fixes #21