feat: add an argument to enable the fetch service (#53)
With this commit the fetch-service service is disabled by default; the way
to enable it is through a new "--use-fetch-service" command line parameter.

This new parameter is initiallly only available to the "pack" command, as
the cache-clearing that the service needs precludes the use of other
lifecycle commands, and the manifest that the application will create based
on the fetch-service's session report is only produced during packing.

Fixes #15