fix: fail managed runs if the build plan is empty
This fix addresses cases like, for example, a project that only builds on `amd64`
being built on `riscv64`. Previously the call to run_managed() would loop over
the (empty) build plan and then finish "successfully". This new version will
raise an EmptyBuildPlanError indicating to the user that they should check the
project's 'platforms' declaration and the command-line parameters.

Fixes #225