feat: support shell/shell-after and debug in 'pack' (#436)
`--shell` and `--shell-after` are implemented for a) consistency with the other
lifecycle commands, and b) because it makes sense to want to inspect the build
environment at packing time, considering that there are packing-related steps,
like writing the metadata file, that the users don't have a way to inspect
otherwise (short of examining the final artefact).

`--debug` is improved to actually shell into the build environment when the
packing itself failed, as opposed to the previous behavior of only debugging
failures that happen during the lifecycle steps.

Fixes #430

---------

Co-authored-by: Alex Lowe <alex.lowe@canonical.com>