fix: properly log the fetch-service's output (#555)
The issue with redirecting the fetch-service's output to a file via bash is
this: as a strict snap, the fetch-service cannot inherit the file descriptors
from the bash process if said process is spawned by a classic snap. This is
the case when the integration is controlled by a snapped craft tool:

```
+----------------------------+        +-------------+
|(snapped) craft-tool -> bash| -----> |fetch-service|
+----------------------------+        +-------------+
```

The solution here is to pass the path to the logfile to the fetch-service itself
through a new, recently implemented command-line option. We also update the
logpath to use tmp, and print the path when emitting the warning about the
integration being experimental.

Refs:

- https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1849753

Fixes #550