refactor: move the test command to lifecycle
The test command in the next lifecycle step after pack, passing the
artifact generated by the pack command to spread tests.

This commit contains a shortcut to pass packed artifact information from
the managed instance back to the outer instance. This must be stored in
a proper location to provide state management for the pack and test
steps of the application lifecycle.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>