feat(application): make the application bootstrap early-run services
A lot of why we were previously putting so much in the Application itself
is because these things needed to occur before we configured the services.

By making service configuration a 2-step process (bootstrap early-run
services at the beginning and then configure later-run services at a
later point), we can offload a lot of this logic to services.