feat(models): make the Platforms schema and processing work the same
This adds a PlatformDict class that behaves as a standard dictionary, but allows pydantic to model it and its schema correctly. The implementation is in a generic class (GenericPlatformsDict) that allows overriding of both the Platform type (if an application needs to modify its Platform model) and the keys that can be used as shorthands.

Fixes #688
CRAFT-4421

Signed-off-by: Alex Lowe <alex.lowe@canonical.com>