feat: add support to spread test expressions
If no arguments are passed to the `test` command, the `craft` backend
is used for testing. If the testing environment is detected as CI,
a system matching the host system is also specified.

If arguments are passed to the `test` command, they are interpreted as
Spread test expressions and passed as-is to the Spread command line
to match job names according to spread rules, including the use of `...`
for globbing.

More information about spread test expressions are available in the
Spread documentation at https://github.com/canonical/spread?#selecting

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>