feat: add support to spread test expressions (#775)
If no arguments are passed to the `test` command, the `craft` backend
is used for testing. If the testing environment is detected as CI,
a system matching the host system is also specified.

If arguments are passed to the `test` command, they are interpreted as
Spread test expressions. If running on CI and test expressions are
specified, filter the list of jobs to match the craft backend and host
system. If running on CI and craft backend is specified, short circuit to
craft backend and host system to avoid unnecessary job expansion.

More information about spread test expressions are available in the
Spread documentation at https://github.com/canonical/spread?#selecting

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>