feat(TestCommand): Pack and test all packable platforms
This changes the test command to pack and test all packable platforms.
Note that it is out of scope as to whether the tests are valid to run.
For example, a cross-compiled riscv64 binary might not be executable if
tested on an amd64 machine, but we will still be running the tests, which
will likely fail.