feat: "support" existing fetch-service sessions
This commit does "half" of the work needed to support packing with pre-existing
sessions that are managed by someone else. It introduces support for the
CRAFT_USE_EXTERNAL_FETCH_SERVICE and CRAFT_PROXY_CERT environment variables and
adds a spread test exposing the current issue: we must configure fetch-service
items early in a managed instance's lifetime.

As such, the feature is not ready for use but this changeset is a step in the
direction of completeness. The remained of the work is described in #839