feat(project): set 26.04 as the current 'devel' base (#955)
This PR bumps the 'devel' Ubuntu base from Questing (25.10) to Resolute (26.04).

It also adds an overridable classmethod so that applications can choose to
have finer control over what is defined as a 'devel' base. Since calling a base
'stable' is a commitment to avoid further changes, applications should be the
'final authority' on what's stable and what's not.