feat(Lifecycle): select plugin groups
This adds the ability to choose plugin groups to the Lifecycle service.

This works differently from the default plugin setup in the Application,
which is now deprecated in favour of this. Apps that move over will be
rewarded with a happy fuzzy feeling and the ability to get the plugin
group based simply on the BuildInfo object.